Add files via upload
Amazon Dataset-Based Recommender System Design Project
Project Overview
This project involves designing and implementing a personalized recommendation system based on Amazon datasets. Amazon is one of the largest e-commerce platforms globally,  with vast customer reviews and product data. In this project, the goal is to analyze this data  using SQL alone and design the basis for a recommendation system that can improve the customer's buying experience.